requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0
opencv-python>=4.8.0
